amp_llm_v3/standalone modules/agent_annotate/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
pydantic-settings
httpx
aiohttp
pyyaml
python-dotenv
python-multipart
